maisb/llm_proxy/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.7
openai>=1.0.0
anthropic>=0.30.0
google-genai>=0.3.0
